pyproject.toml
[tool.poetry]
name = "project-metrics"
version = "0.0.1"
description = "Overview of projects and their health"
authors = ["Gustav Rasmussen"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.isort]
profile = "black"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pygithub = "^2.4.0"
pandas = "^2.2.3"
tabulate = "^0.9.0"
numpy = "2.2.0"
sphinx = "^8.1.3"
seaborn = "^0.13.2"
icecream = "^2.1.3"
loguru = "^0.7.2"
file-convertion-tools = "^0.0.8"
dynaconf = "^3.2.7"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
mypy = "^1.14.0"
pre-commit = "^3.8.0"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
ruff = "^0.9.2"
wily = "^1.24.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
